requirements.txt
torch
diffusers[torch]
Pillow
vtracer
cairosvg
click
transformers
scikit-learn
torchvision
